refactored SqlRenameTable by adding base class SqlRename
