+++ parsing/dialects/dialect1/config.fmpp
@@ -81,8 +81,9 @@ data: {
       "org.apache.calcite.sql.SqlPeriodTypeNameSpec",
       "org.apache.calcite.sql.SqlPeriodTypeNameSpec.TimeScale",
       "org.apache.calcite.sql.SqlPrimaryIndex",
+      "org.apache.calcite.sql.SqlRename",
+      "org.apache.calcite.sql.SqlRename.RenameOption",
       "org.apache.calcite.sql.SqlRenameTable",
-      "org.apache.calcite.sql.SqlRenameTable.RenameOption",
       "org.apache.calcite.sql.SqlSecondaryIndex",
       "org.apache.calcite.sql.SqlSelectTopN",
       "org.apache.calcite.sql.SqlSetTimeZone",
