+++ parsing/dialects/dialect1/parserImpls.ftl
@@ -1456,7 +1456,7 @@ SqlDataTypeSpec DataTypeAlternativeCastSyntax() :
     }
 }
 
-SqlRenameTable SqlRenameTable() :
+SqlRename SqlRenameTable() :
 {
     SqlIdentifier targetTable;
     SqlIdentifier sourceTable;
